chore: try to avoid swt jni loading issue in commandline
