+++ main/plugins/org.talend.sdk.component.studio-integration/src/main/java/org/talend/sdk/component/studio/Lookups.java
@@ -26,6 +26,7 @@ import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Composite;
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceReference;
+import org.talend.commons.CommonsPlugin;
 import org.talend.core.GlobalServiceRegister;
 import org.talend.core.IService;
 import org.talend.core.model.components.IComponent;
@@ -52,6 +53,11 @@ public final class Lookups {
     }
 
     public static Runnable init() {
+        if (CommonsPlugin.isHeadless()) {
+            // currently the following service proxy only used for GUI part, ignore it in cmd
+            return () -> {
+            };
+        }
         try {
             final Field instance = GlobalServiceRegister.class.getDeclaredField("instance");
             if (!instance.isAccessible()) {
